note to self: make git use vim instead of nano, since that is what you are used to. second note to self: remember that you just learned about :e :split :vsplit and C-w commands in vim, they seem very useful.
This commit fixes all really huge image sizes, resizing them down to be a percentage of the parent window. This was accomplished in markdown using {:width="33%"} after the image ![]() tag. in order to do this most efficiently (and to brush up on my sed), I used the sed command:

sed -i.bak 's/\(!\[[^(]*\]([^\[]*)\)[^{]/\1{:width="25%"} /g' *.md
this looks for ![stuff](stuff) and makes it into ![stuff](stuff}{:width="25%"}. The basic sed syntax to do that would be
sed 's/\(!\[.*\](.*)\)/\1{:width="25%"}/g'
(note that parentheses need to be escaped in order to capture, whereas brackets need to be escaped in order to be literal)
HOWEVER, as I immediately found out, since sed does greedy matching, I had to add some syntax to make it less greedy. Instead of looking for .* between brackets, it looks for [^(]*, meaning the largest contiguous glob of NON-( characters.
I also make sure to only do matches that don't have the curly braces already, by including a [^{] specifier at the end.